+++ src/HSMP/HSMPResponse.hpp
@@ -99,7 +99,7 @@ struct BroadcastResponse : ServerResponse {
 
 struct UploadFileResponse : ServerResponse {
   short int tam_file_name;
-  short int tam_file_data;
+  int tam_file_data;
   short int tam_remitente;
   std::string file_name;
   std::string file_data;
